backend/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
sqlalchemy==1.4.27
databases==0.5.4
python-jose==3.3.0
python-multipart==0.0.5
aiosqlite==0.17.0
passlib
argon2_cffi
shortuuid